Feature Abrir Desenho

+++ Paint/src/editor/componentes/listeners/SalvarDesenho.java
@@ -8,6 +8,7 @@ import java.util.Iterator;
 import java.util.Vector;
 
 import javax.swing.JFileChooser;
+import javax.swing.JOptionPane;
 import javax.swing.filechooser.FileNameExtensionFilter;
 
 import editor.Janela;
@@ -28,20 +29,28 @@ public class SalvarDesenho extends EditorActionListener {
             int returnVal = fc.showSaveDialog(null);
             if (returnVal == JFileChooser.APPROVE_OPTION) {
                 File file = fc.getSelectedFile();
-                BufferedWriter escritor = new BufferedWriter(new FileWriter(file));
-                Vector<Figura> figuras = this.janela.getPainelDesenho().getFiguras();
 
-                Iterator figura = figuras.iterator();
-                while (figura.hasNext()) { 
-                    escritor.write(figura.next().toString());
-                    escritor.newLine();
+                int input = 0;
+                if (file.exists()) 
+                    input = JOptionPane.showConfirmDialog(null, "Tem certeza que quer sobreescrever este arquivo?");
+                
+                if (input == 0) {
+                    BufferedWriter escritor = new BufferedWriter(new FileWriter(file));
+                    Vector<Figura> figuras = this.janela.getPainelDesenho().getFiguras();
+
+                    Iterator figura = figuras.iterator();
+                    while (figura.hasNext()) { 
+                        escritor.write(figura.next().toString());
+                        escritor.newLine();
+                    }
+
+                    escritor.close();
                 }
-
-                escritor.close();
             }
         }
         catch(Exception ex) {
-
+            JOptionPane.showMessageDialog(null, "Erro na abertura do arquivo:", "",
+                    JOptionPane.WARNING_MESSAGE);
         }
 
     }
